api/requirements.txt
annotated-types
anyio
arrow
binaryornot
certifi
chardet
charset-normalizer
click
colorama
cookiecutter
dnspython
email_validator
fastapi
fastapi-cli
gunicorn
h11
httpcore
httptools
httpx
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
oauthlib
packaging
pydantic
pydantic_core
Pygments
pypng
python-dateutil
python-dotenv
python-multipart
python-slugify
PyJWT
PyYAML
qrcode
repath
requests
rich
rich-toolkit
shellingham
six
sniffio
starlette
text-unidecode
toml
typer
types-python-dateutil
typing-inspection
typing_extensions
urllib3
uvicorn
watchdog
watchfiles
websockets
python-jose[cryptography]
